feat: renamed Modal to Popup. feat: creating SettingsPopup. - when using PopupCard, always specify a width.
