+++ frontend/src/views/PopupTest.tsx
@@ -5,19 +5,15 @@ import Logo from "../components/Logo";
 import Input from "../components/Input";
 import Button from "../components/Button";
 import Divider from "../components/Divider";
-import Popup from "../popups/Popup";
+import Popup from "../popups/SettingsPopup";
 
 const PopupTest: React.FC = () => {
   const [showPopup, setShowPopup] = useState(false);
 
   return (
     <Background>
-      <Card>
-        <Logo />
-        <Input placeholder="Username" />
-        <Input placeholder="Password" type="password" />
+      <Card className="flex justify-center">
         <Button onClick={() => setShowPopup(true)}>LOGIN</Button>
-        <Divider />
       </Card>
       <Popup open={showPopup} onClose={() => setShowPopup(false)} />
     </Background>
